+++ app/src/main/java/org/breezyweather/common/ui/composables/LocationPreferences.kt
@@ -19,7 +19,6 @@ package org.breezyweather.common.ui.composables
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll