+++ app/src/main/java/org/breezyweather/settings/activities/SettingsActivity.kt
@@ -42,7 +42,6 @@ import breezyweather.domain.location.model.Location
 import com.google.accompanist.permissions.rememberPermissionState
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
-import org.breezyweather.BuildConfig
 import org.breezyweather.R
 import org.breezyweather.common.basic.GeoActivity
 import org.breezyweather.common.bus.EventBus