+++ app/src/src_nofdroid/org/breezyweather/sources/ims/json/ImsWindDirection.kt
@@ -16,7 +16,6 @@
 
 package org.breezyweather.sources.ims.json
 
-import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable